function/requirements.txt
google-cloud-tasks==2.7.1